pyproject.toml
[tool.poetry]
name = "universal-scrapper"
version = "0.0.1"
description = "universal scrapper for all websites"
authors = ["Jayaraj Esvar <jayaraj.esvar@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.1"
openai = "^1.30.1"
firecrawl-py = "^0.0.9"
tenacity = "^8.3.0"
termcolor = "^2.4.0"
tiktoken = "^0.7.0"
agentql = "^0.4.0"
playwright = "^1.44.0"
groq = "^0.8.0"

[toot.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
#https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
